refactor(client/filesystem): move into internal/client/filesystem subpackage (#140)
The 1233-line internal/client/filesystem_client.go now lives under
internal/client/filesystem/, split per CRD domain:

  client.go         core type + controller-runtime Client interface
  mcpserver.go      MCPServer CRUD
  serviceclass.go   ServiceClass CRUD
  workflow.go       Workflow CRUD
  events.go         Kubernetes Event API

The MusterClient interface stays in the parent client package. The only
caller of the old NewFilesystemClient was the dispatcher in
muster_client.go itself, which now calls filesystem.New(basePath)
directly. External consumers depend only on MusterClient and are
unaffected.

filesystemClient                                    -> filesystem.Client
NewFilesystemClient                                 -> filesystem.New(basePath string) *Client
filesystemStatusWriter, filesystemSubResourceClient -> unexported
