+++ internal/client/muster_client.go
@@ -11,6 +11,7 @@ import (
 	musterv1alpha1 "github.com/giantswarm/muster/pkg/apis/muster/v1alpha1"
 
 	"github.com/giantswarm/muster/internal/api"
+	"github.com/giantswarm/muster/internal/client/filesystem"
 	"github.com/giantswarm/muster/pkg/logging"
 )
 
@@ -110,7 +111,7 @@ func NewMusterClientWithConfig(cfg *MusterClientConfig) (MusterClient, error) {
 	}
 
 	// Fall back to filesystem mode
-	return NewFilesystemClient(cfg)
+	return filesystem.New(cfg.FilesystemPath), nil
 }
 
 // MusterClientConfig provides configuration options for client creation.
